Fixed isses raised and missing files
