+++ tests/UI Testing/TileSets/TestTileset.tms
