+++ tests/UI Testing/UI Testing.csproj
@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <ItemGroup>
-    <ProjectReference Include="..\Tile-Mapper.csproj" />
+    <ProjectReference Include="..\..\src\Tile-Mapper.csproj" />
   </ItemGroup>
 
   <PropertyGroup>
